refactor(llmproviders): share the common Google wire types
GoogleGenerativeAI and GoogleGeminiCli each carried byte-identical
copies of sanitize_schema, schema, and the generateContent wire structs
(FunctionDeclaration through Candidate). Move those into a new
GoogleShared module that both provider modules pull in via
`using ..GoogleShared: ...`, so GoogleGenerativeAI.X and
GoogleGeminiCli.X keep resolving for all downstream references in the
Agentif adapters and tests. UsageMetadata is kept as the superset of
both variants (the generative side simply leaves thoughtsTokenCount and
cachedContentTokenCount as nothing), matching how the two usage
functions already read it. The gemini-cli module's `Model` import stays:
build_request reads model.id, and rethreading a bare id through its
stream.jl call site would churn more than it saves.
